pyproject.toml
[project]
name = "learning-ai"
version = "0.1.0"
description = "AI microservice for lesson generation, quiz generation, and coaching"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "gunicorn>=23",
  "pydantic>=2.9",
  "httpx>=0.27",
  "google-cloud-aiplatform>=1.70",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "pytest-asyncio>=0.24",
  "httpx>=0.27",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.uv]
package = false
